fix: solve disvocered issues by using test cases

+++ src/FlexLexer.h
@@ -60,7 +60,7 @@ class FlexLexer {
   virtual ~FlexLexer() {}
 
   const char* YYText() const { return yytext; }
-  size_t YYLeng() const { return yyleng; }
+  int YYLeng() const { return yyleng; }
 
   virtual void yy_switch_to_buffer(yy_buffer_state* new_buffer) = 0;
   virtual yy_buffer_state* yy_create_buffer(std::istream* s, int size) = 0;
@@ -94,7 +94,7 @@ class FlexLexer {
 
  protected:
   char* yytext;
-  size_t yyleng;
+  int yyleng;
   int yylineno;       // only maintained if you use %option yylineno
   int yy_flex_debug;  // only has effect with -d or "%option debug"
 };
@@ -139,8 +139,8 @@ class yyFlexLexer : public FlexLexer {
   virtual int yywrap();
 
  protected:
-  virtual size_t LexerInput(char* buf, size_t max_size);
-  virtual void LexerOutput(const char* buf, size_t size);
+  virtual int LexerInput(char* buf, int max_size);
+  virtual void LexerOutput(const char* buf, int size);
   virtual void LexerError(const char* msg);
 
   void yyunput(int c, char* buf_ptr);
@@ -169,7 +169,7 @@ class yyFlexLexer : public FlexLexer {
   char yy_hold_char;
 
   // Number of characters read into yy_ch_buf.
-  size_t yy_n_chars;
+  int yy_n_chars;
 
   // Points to current character in buffer.
   char* yy_c_buf_p;
